# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Hotel Management System API",
        "description": "API documentation for the Hotel Management System",
        "contact": {
            "name": "Support Team",
            "email": "support@hotelsystem.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://my-default-host.com",
            "description": "Hotel Management System API Server"
        }
    ],
    "paths": {
...
        "/api/v1/payments/create-intent": {
            "post": {
                "tags": [
                    "Payments"
                ],
                "summary": "Create a payment intent",
                "operationId": "f2faf2a637cb6235b4eac10c5653d0a7",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "booking_id"
                                ],
                                "properties": {
                                    "booking_id": {
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Payment intent created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "clientSecret": {
                                            "type": "string"
                                        },
                                        "paymentIntentId": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/v1/payments/confirm": {
            "post": {
                "tags": [
                    "Payments"
                ],
                "summary": "Confirm a payment",
                "operationId": "44fe27737fbdf48934f11b59ace7a4b9",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "payment_intent_id"
                                ],
                                "properties": {
                                    "payment_intent_id": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Payment confirmed successfully"
                    }
                }
            }
        },
        "/api/v1/payments/{payment}/refund": {
            "post": {
                "tags": [
                    "Payments"
                ],
                "summary": "Refund a payment",
                "operationId": "0436543574a2161068cbbf25840912fd",
                "parameters": [
                    {
                        "name": "payment",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "amount": {
                                        "type": "number"
                                    },
                                    "reason": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Payment refunded successfully"
                    }
                }
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            },
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            }
        }
    },
...
    "security": [
        {
            "bearerAuth": []
        }
    ]
}